fix issue from comment
